rename library to describe what is actually taking place: we are making liquidity!
